s1singlescatter update
